+++ src/JuMP.jl
@@ -31,7 +31,7 @@ export
     getInternalModel, buildInternalModel,
     # Variable
     setName, getName, setLower, setUpper, getLower, getUpper,
-    getValue, setValue, getDual, setVarType,
+    getValue, setValue, getDual, setCategory, getCategory, 
     # Expressions and constraints
     affToStr, quadToStr, conToStr, chgConstrRHS,
     
@@ -245,10 +245,12 @@ function getDual(v::Variable)
 end
 
 const var_types = [:Cont, :Int, :Bin, :SemiCont, :SemiInt]
-function setVarType(v::Variable, v_type::Symbol)
+function setCategory(v::Variable, v_type::Symbol)
     v_type in var_types || error("Unrecognized variable type $v_type. Should be one of:\n    $var_types")
     v.m.colCat[v.col] = v_type
 end
+
+getCategory(v::Variable) = v.m.colCat[v.col]
 
 Base.zero(v::Type{Variable}) = AffExpr(Variable[],Float64[],0.0)
 Base.zero(v::Variable) = zero(typeof(v))